% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcstats.R
\docType{package}
\name{arcstats-package}
\alias{arcstats}
\alias{arcstats-package}
\title{arcstats: processing and physical activity summaries of minute-level activity data}
\description{
arcstats provides `activity_stats()` function to process minute level
actigraphy-measured activity counts
data and extract commonly used physical activity volume and fragmentation metrics.
}
\details{
To learn more about arcstats, start with the vignettes:
`browseVignettes(package = "arcstats")`
}
\author{
\strong{Maintainer}: Marta Karas \email{marta.karass@gmail.com} (\href{https://orcid.org/0000-0001-5889-3970}{ORCID})

Authors:
\itemize{
  \item Jacek Urbanek (\href{https://orcid.org/0000-0002-1890-8899}{ORCID})
  \item Jennifer Schrack
}

}
